Modification to imcmp app by adding command line options for conformance testing.
imgcmp app is to compare two images in PAE (Peak Absolute Error) and MSE (Mean Squared Error). Additionally, it shows PSNR (Peak-Signal-to-Noise-Ratio).
This commit adds a functionality to determine the PAE and MSE values are less than or equal to the allowable values given as the last two command line options.